+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  + Add new command line option.
    + “-r” can now take a number to start from.
    + “-l” list commands of the configuration file with its number.
+   + “-a” add the command given on stdin to configuration file.
  + Improve some messages.
  + Display run commands in title bar of the windows terminal.
  + New tmp file